♻️ refactor(generation): move the topic url sync into src/features
The route tree is for thin page-segment composition; a store/URL synchronisation
unit is domain logic and belongs under src/features. GenerationLayout now only
composes it, and the image/video route layouts only pass their store.
