+++ src/features/Generation/index.ts
@@ -1,0 +1,2 @@
+export { default as TopicUrlSync } from './TopicUrlSync';
+export { type TopicUrlSyncStore, useTopicUrlSync } from './useTopicUrlSync';
